requirements.txt
python-escpos
